LRCI-5973 Create modules batch build testray case results into one class
