+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayFactory.java
@@ -14,8 +14,8 @@ import com.liferay.jenkins.results.parser.test.clazz.group.AxisTestClassGroup;
 import com.liferay.jenkins.results.parser.test.clazz.group.FunctionalAxisTestClassGroup;
 import com.liferay.jenkins.results.parser.test.clazz.group.JSUnitAxisTestClassGroup;
 import com.liferay.jenkins.results.parser.test.clazz.group.JUnitAxisTestClassGroup;
+import com.liferay.jenkins.results.parser.test.clazz.group.ModulesAxisTestClassGroup;
 import com.liferay.jenkins.results.parser.test.clazz.group.PlaywrightAxisTestClassGroup;
-import com.liferay.jenkins.results.parser.test.clazz.group.SemVerModulesAxisTestClassGroup;
 
 import java.io.File;
 
@@ -93,10 +93,8 @@ public class TestrayFactory {
 					testrayBuild, topLevelBuildReport, axisTestClassGroup,
 					testClass, testClassMethod);
 			}
-			else if (axisTestClassGroup instanceof
-						SemVerModulesAxisTestClassGroup) {
-
-				return new SemVerModulesBatchBuildTestrayCaseResult(
+			else if (axisTestClassGroup instanceof ModulesAxisTestClassGroup) {
+				return new ModulesBatchBuildTestrayCaseResult(
 					testrayBuild, topLevelBuildReport, axisTestClassGroup,
 					testClass);
 			}
